View the factors with the most extreme averages for each field

View the factors with the most extreme averages for each field

summarize_factor_extremes(...)

Arguments

...
Arguments passed on to refactor_columns

df

dataframe to evaluate

dep_var

dependent variable to use (column name)

split_on

variable to split data / group by

id

field to use as ID

n_cat

for categorical variables, the max number of unique values to keep. This field feeds the forcats::fct_lump(n = ) argument.

collapse_by

should n_cat collapse by the distance to the grand mean "dv" leaving the extremes as is and grouping factors closer to the grand mean as "other" or should it use size "n"

n_quantile

for numeric/date fields, the number of quantiles used to split the data into a factor. Fields that have less than this amount will not be changed.

n_digits

for numeric fields, the number of digits to keep in the breaks ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if n_digits = 2

avg_type

mean or median

ignore_cols

columns to ignore from analysis. Good candidates are fields that have have no duplicate values (primary keys) or fields with a large proportion of null values

Examples

summarize_factor_extremes(df = ggplot2::mpg, dep_var = "hwy")

#> Joining, by = "field"

#> # A tibble: 10 x 9
#>    adj_r_sq   p_value n_low value_low  avg_low field  avg_high value_high n_high
#>       <dbl>     <dbl> <int> <chr>        <dbl> <fct>     <dbl> <chr>       <int>
#>  1  0.906   1.11e-112    25 01 [8.97 ~    14.8 cty        31.2 06 [22 to~     12
#>  2  0.680   1.26e- 54    33 pickup        16.9 class      28.3 compact        47
#>  3  0.669   7.60e- 51    29 06 [4.3 t~    16.9 displ      30.8 01 [1.59 ~     43
#>  4  0.579   1.18e- 45    70 8             17.6 cyl        28.8 4              81
#>  5  0.527   1.13e- 38   103 4             19.2 drv        28.2 f             106
#>  6  0.484   3.25e- 29    10 ram 1500 ~    15.3 model      32.8 new beetle      6
#>  7  0.483   3.67e- 29     8 jeep          17.6 manuf~     32.6 honda           9
#>  8  0.192   7.93e- 11     8 e             13.2 fl         25.2 p              52
#>  9  0.113   3.49e-  5     6 auto(l6)      20   trans      26.3 manual(m5)     58
#> 10 -0.00431 9.74e-  1   117 1999          23.4 year       23.5 2008          117