Summarize fields

Pivots data and summarizes factor frequencies by field and generates stats used for plotting

analyze_data(
  df,
  ...,
  avg_type = c("mean", "median"),
  return = c("data", "list")
)

Arguments

df
dataframe to evaluate

...
Arguments passed on to refactor_columns

dep_var

dependent variable to use (column name)

n_cat

for categorical variables, the max number of unique values to keep. This field feeds the forcats::fct_lump(n = ) argument.

n_quantile

for numeric/date fields, the number of quantiles used to split the data into a factor. Fields that have less than this amount will not be changed.

n_digits

for numeric fields, the number of digits to keep in the breaks ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if n_digits = 2

ignore_cols

columns to ignore from analysis. Good candidates are fields that have have no duplicate values (primary keys) or fields with a large proportion of null values

avg_type
mean or median

return
option to return a dataframe when TRUE and a list when FALSE. The list option includes the original min/max of the data and the grand average.

See also

Examples

analyze_data(iris, dep_var = "Sepal.Length")

#> Joining, by = "field"

#> # A tibble: 26 x 10
#>    field value factor_avg     n field_r_sq field_r_sq_adj field_p_value
#>    <fct> <chr>      <dbl> <int>      <dbl>          <dbl>         <dbl>
#>  1 Sepa~ 02 [~       5.31     7      0.110         0.0524      5.44e- 2
#>  2 Sepa~ 03 [~       5.89    22      0.110         0.0524      5.44e- 2
#>  3 Sepa~ 04 [~       6.22    24      0.110         0.0524      5.44e- 2
#>  4 Sepa~ 05 [~       6.02    37      0.110         0.0524      5.44e- 2
#>  5 Sepa~ 06 [~       5.69    31      0.110         0.0524      5.44e- 2
#>  6 Sepa~ 07 [~       5.26    10      0.110         0.0524      5.44e- 2
#>  7 Sepa~ 08 [~       5.75    11      0.110         0.0524      5.44e- 2
#>  8 Peta~ 01 [~       4.98    37      0.827         0.817       7.18e-50
#>  9 Peta~ 02 [~       5.07    13      0.827         0.817       7.18e-50
#> 10 Peta~ 05 [~       5.49     8      0.827         0.817       7.18e-50
#> # ... with 16 more rows, and 3 more variables: rescale_factor_avg <dbl>,
#> #   grand_avg <dbl>, field_wt <dbl>

analyze_data(iris, dep_var = "Sepal.Length", return = "list")

#> Joining, by = "field"

#> $data
#> # A tibble: 26 x 10
#>    field value factor_avg     n field_r_sq field_r_sq_adj field_p_value
#>    <fct> <chr>      <dbl> <int>      <dbl>          <dbl>         <dbl>
#>  1 Sepa~ 02 [~       5.31     7      0.110         0.0524      5.44e- 2
#>  2 Sepa~ 03 [~       5.89    22      0.110         0.0524      5.44e- 2
#>  3 Sepa~ 04 [~       6.22    24      0.110         0.0524      5.44e- 2
#>  4 Sepa~ 05 [~       6.02    37      0.110         0.0524      5.44e- 2
#>  5 Sepa~ 06 [~       5.69    31      0.110         0.0524      5.44e- 2
#>  6 Sepa~ 07 [~       5.26    10      0.110         0.0524      5.44e- 2
#>  7 Sepa~ 08 [~       5.75    11      0.110         0.0524      5.44e- 2
#>  8 Peta~ 01 [~       4.98    37      0.827         0.817       7.18e-50
#>  9 Peta~ 02 [~       5.07    13      0.827         0.817       7.18e-50
#> 10 Peta~ 05 [~       5.49     8      0.827         0.817       7.18e-50
#> # ... with 16 more rows, and 3 more variables: rescale_factor_avg <dbl>,
#> #   grand_avg <dbl>, field_wt <dbl>
#> 
#> $field_stats
#> # A tibble: 4 x 4
#>   field        field_r_sq field_r_sq_adj field_p_value
#>   <chr>             <dbl>          <dbl>         <dbl>
#> 1 Sepal.Width       0.110         0.0524      5.44e- 2
#> 2 Petal.Length      0.827         0.817       7.18e-50
#> 3 Petal.Width       0.685         0.665       5.95e-31
#> 4 Species           0.619         0.614       1.67e-31
#> 
#> $factor_stats
#> # A tibble: 26 x 4
#>    field        value             factor_avg     n
#>    <chr>        <chr>                  <dbl> <int>
#>  1 Sepal.Width  02 [2.24 to 2.48)       5.31     7
#>  2 Sepal.Width  03 [2.48 to 2.72)       5.89    22
#>  3 Sepal.Width  04 [2.72 to 2.96)       6.22    24
#>  4 Sepal.Width  05 [2.96 to 3.2)        6.02    37
#>  5 Sepal.Width  06 [3.2 to 3.44)        5.69    31
#>  6 Sepal.Width  07 [3.44 to 3.68)       5.26    10
#>  7 Sepal.Width  08 [3.68 to 3.92)       5.75    11
#>  8 Petal.Length 01 [0.99 to 1.59)       4.98    37
#>  9 Petal.Length 02 [1.59 to 2.18)       5.07    13
#> 10 Petal.Length 05 [3.36 to 3.95)       5.49     8
#> # ... with 16 more rows
#> 
#> $grand_avg
#> [1] 5.843333
#> 
#> $orig_min
#>  2% 
#> 4.4 
#> 
#> $orig_max
#> 98% 
#> 7.7 
#>